requirements.txt
stix
stix-validator
stix2
stix2-validator
stix2-patterns==1.3.2
carbon-black-cloud-sdk==1.3.5
validators
cabby
taxii2-client
PyYAML

# Dev dependencies
pytest==6.2.5
coverage==6.3
coveralls==3.3.1
flake8==4.0.1
flake8-colors==0.1.9
flake8-docstrings==1.6.0
